refactor: 뷰어/에디터 경로 정리
